completed)

+++ Project/Contacts/src/main/java/com/example/Contacts/repos/ContactsRepo.java
@@ -1,5 +1,5 @@
 package com.example.Contacts.repos;
-import com.example.Contacts.domain.ContactsList;
+import com.example.Contacts.domain.dto.ContactsList;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
@@ -7,6 +7,6 @@ import java.util.List;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface ContactsRepo extends CrudRepository<ContactsList, Integer> {
+public interface ContactsRepo extends CrudRepository<ContactsList, Long> {
     List<ContactsList> findByContactName(String name);
 }
